==> 01. meta_rot_proc_ksc_241127.ipynb <==
import pandas as pd
from tqdm import tqdm

tqdm.pandas()

meta_movie = pd.read_pickle('/data/meta/241122_metadata_us_movie.pickle')
meta_movie = pd.DataFrame(meta_movie)
meta_movie['domain'] = 'movie'
meta_movie.head(2)

meta_tv = pd.read_pickle('/data/meta/241122_metadata_us_tvshow.pickle')
meta_tv = pd.DataFrame(meta_tv)
meta_tv['domain'] = 'tv_show'
meta_tv.head(2)

movie = meta_movie[['contents_set_id', 'contents_id', 'execution_info', 'domain', 'description', 'thumbnail_info', 'actor_info', 'director_info', 'cntry_code', 'document_id', 'title_name','genre_info']]

print(movie.shape)
sum(movie.execution_info.isna())

meta_tv.columns

meta_tv.tail(2)

meta_movie.tail(2)

tv = meta_tv[['contents_set_id', 'contents_id', 'execution_info', 'description', 'thumbnail_info', 'actor_info', 'director_info', 'cntry_code', 'document_id', 'title_name','genre_info', 'domain']]
tv.tail(2)

print(tv.shape)
sum(tv.execution_info.isna())

meta = pd.concat([movie, tv], ignore_index=True)
print(meta.shape)
meta.tail(2)

import re

korean = re.compile("[가-힣]+")
meta["genre_info"] = [korean.findall(str(i)) for i in meta["genre_info"]]

# actor -> 배우 이름 / 극중 이름으로 기재되어 있어서 배우 이름만 발라냄
meta["actor_info"] = meta["actor_info"].apply(lambda x : str(x).replace("\x0b", "").replace("\x0c", ""))

actor = re.compile("[0-9].+")
non_num = re.compile("[0-9]+")

res = []
for i in tqdm(meta["actor_info"]):
    temp = i.split("Actor")
    actors = []
    for w in temp:
        try:
            actors.append(re.sub(non_num, "", actor.findall(w)[0]))
        except:
            pass
    res.append(actors)
meta["actor_info"] = res

meta['new_content_id'] = meta.index
meta.tail(2)

meta[meta['contents_id'].str.contains('TMS')]['execution_info'].iloc[0]

meta[meta['execution_info'].notnull()].to_csv('./data/Meta_241127.csv')
