==> 02. search_log_proc_241127.ipynb <==
import pandas as pd
import numpy as np

search = pd.read_csv('/data/log-data-2024/search_log_09_11.csv')
search

search['log_time'] = pd.to_datetime(search['log_time'])

# api 컬럼에서 logging인 데이터만 추출
search['api_logging'] = search['api'].apply(lambda x : x.split('/')[-1])
search = search[search['api_logging'] == 'logging'].reset_index(drop=True)

# domain이 tvshow, movie인 데이터만 추출
search['domain'] = search['params'].apply(lambda x: next((part.split('=')[1] for part in x.split('&') if part.startswith('domain=')), None))
search.head(2)

search = search[search['domain'].isin(['movie', 'tvshow'])].reset_index(drop=True)

# params에서 content_id, content_set_id 추출
search['content_id'] = search['params'].apply(lambda x: next((part.split('=')[1] for part in x.split('&') if part.startswith('content_id=')), None))
search['content_set_id'] = search['params'].apply(lambda x: next((part.split('=')[1] for part in x.split('&') if part.startswith('content_set_id=')), None))

search = search[['device_id', 'log_time', 'content_id', 'content_set_id', 'api_logging']]

search.to_csv('./data/search_log_241127.csv', index=False)

search.tail()
