==> 03. sequence_generate_search_241127.ipynb <==
import pandas as pd
import numpy as np
import re
import ast
import uuid
from datetime import datetime, timedelta
from collections import Counter
from tqdm import tqdm

tqdm.pandas()

meta_path = '/data/log-data-2024/1.meta_rot_preprocessing_ksc/data/Meta_241127.csv'
search_log_path = '/data/log-data-2024/1.meta_rot_preprocessing_ksc/data/search_log_241127.csv'
recommend_path = '/data/log-data-2024/recommend_click_log_all_09_11.csv'
sequence_output_path = '/data/log-data-2024/2.sequence_generate_ksc/data/input_241127.txt'
sequence_device_path = '/data/log-data-2024/2.sequence_generate_ksc/data/sequence_device_match_241127.csv'
title_id_match_path = '/data/log-data-2024/2.sequence_generate_ksc//data/title_id_match_241127.csv'

meta = pd.read_csv(meta_path, index_col=0)
meta.reset_index(drop=True, inplace=True)
print(meta.shape)

meta = meta.rename(columns = {'contents_id': 'content_id', 'contents_set_id':'content_set_id', 'new_contents_id':'new_content_id'})

meta.head(2)

app_id = pd.read_csv('/data/log-data-2024/appId_setId_mapping.csv')
app_id.head(2)

recommend_click = pd.read_csv(recommend_path)
recommend_click = pd.merge(recommend_click, app_id, left_on = 'message_value_1', right_on='app_id', how='inner')

recommend_click = recommend_click[['device_id', 'log_time', 'message_value_1', 'message_value_2', 'message_value_4', 'contents_set_id']]

recommend_click = recommend_click.rename(columns = {'message_value_1': 'app_id', 'message_value_2': 'content_id', 'contents_set_id':'content_set_id'})

recommend_click = pd.merge(recommend_click, meta[['content_id', 'content_set_id', 'new_content_id']], on=['content_id', 'content_set_id'], how='left')


print(recommend_click.shape)
recommend_click.tail(2)

recommend_click[recommend_click['new_content_id'].isnull()] 

meta[meta['content_id'] == 'amzn1.dv.gti.704dcd96-2364-4ddd-b10e-f3dac081aef9']

meta[meta['title_name']== 'Sonic the Hedgehog 2'][['content_set_id', 'content_id', 'execution_info', 'title_name']]



meta[meta['title_name']== 'Sonic the Hedgehog 2'].iloc[1].execution_info

recommend_click = recommend_click[recommend_click['new_content_id'].notnull()]

# recommend_click = recommend_click[(recommend_click['log_time'] >= '2024-08-01') & (recommend_click['log_time'] < '2024-09-01')]

search = pd.read_csv(search_log_path)
search = pd.merge(search, meta[['content_id', 'content_set_id', 'new_content_id']], on=['content_id', 'content_set_id'], how='left')
print(search.shape)
search.tail(2)

def data_preprocessing(recommend_click):
    # 정렬
    recommend_click = recommend_click.sort_values(by=['device_id', 'log_time'])
    
    # 이상치 제거
    recommend_click['log_time'] = recommend_click['log_time'].progress_apply(lambda x: x.split('.')[0])
    recommend_click['log_time'] = pd.to_datetime(recommend_click['log_time'], format='%Y-%m-%d %H:%M:%S')
    
    # log_time을 datetime으로 변환
    recommend_click['log_time'] = pd.to_datetime(recommend_click['log_time'])

    return recommend_click



def data_preprocessing_search(search):
    # 정렬
    search = search.sort_values(by=['device_id', 'log_time'])
    # 이상치 제거
    search['log_time'] = search['log_time'].progress_apply(lambda x: x.split('.')[0])
    search['log_time'] = pd.to_datetime(search['log_time'], format='%Y-%m-%d %H:%M:%S')

    # search_df = search[['device_id', 'message_value_4', 'log_time']].copy()
    
    # log_time을 datetime으로 변환
    search['log_time'] = pd.to_datetime(search['log_time'])
    
    return search



def generate_sequences(df, col_name, day_term=30):
    time_term = timedelta(days=day_term)
    # df['log_time'] = pd.to_datetime(df['log_time'])
    df = df.sort_values(by='log_time')
    sequences = []
    current_sequence = []
    previous_time = None
    
    for _, row in df.iterrows():
        if previous_time is None or (row['log_time'] - previous_time) >= time_term:
            if len(current_sequence) > 2:
                sequences.append(current_sequence)
            current_sequence = [row[col_name]]
        else:
            current_sequence.append(row[col_name])
        previous_time = row['log_time']
    
    if len(current_sequence) > 2: 
        sequences.append(current_sequence)
    
    return sequences


def convert_list_to_txt(click_df, col_name, day_term=30):
    # 각 numeric_device_id에 대해 시퀀스를 생성하고 리스트로 변환
    grouped = click_df.groupby('device_id')
    
    total_sequences = []
    sequence_device = {}  # device_id와 시퀀스 인덱스 매핑을 위한 딕셔너리
    idx = 0  # 시퀀스 인덱스 초기화

    for device_id, group in grouped:
        # device_id에 대한 시퀀스를 생성
        sequences = generate_sequences(group, col_name, day_term)
        total_sequences.extend(sequences)
        
        # 각 시퀀스에 대해 시퀀스 번호와 device_id 매핑 저장
        for _ in sequences:
            sequence_device[idx] = device_id
            idx += 1
    
    # 결과를 문자열로 변환하여 텍스트 파일에 저장
    result = []
    for idx, item in enumerate(total_sequences):
        content_string = str(idx) + ' ' + ' '.join(map(str, item)) + '\n'
        result.append(content_string)

    # 시퀀스와 device_id 매핑을 DataFrame으로 변환하고 CSV로 저장
    sequence_device_df = pd.DataFrame(list(sequence_device.items()), columns=['seq_idx', 'device_id'])
    sequence_device_df.to_csv(sequence_device_path, index=False)
    
    # 시퀀스 데이터를 텍스트 파일로 저장
    with open(sequence_output_path, 'w') as file:
        file.writelines(result)




def main(recommend_click, search, day_term=30):
    print('--------recommend_click 전처리--------')
    rec_temp = data_preprocessing(recommend_click)

    print('-------------search 전처리-------------')
    search_temp = data_preprocessing_search(search)

    temp = pd.concat([rec_temp, search_temp], axis=0)

    print('---recommend_click, search 결합 완료---')

    temp['log_time'] = pd.to_datetime(temp['log_time'], format='%Y-%m-%d %H:%M:%S.%f')
    result = temp[temp['new_content_id'].notna()]
    
    print('-----------list 파일로 변환------------')
    convert_list_to_txt(result, 'new_content_id', day_term)

    return result

result = main(recommend_click, search, 30)
result.head(3)

unique_result = result[['message_value_4', 'contents_id']]
unique_result.to_csv(title_id_match_path, index=False)

### sequence 확인

import pandas as pd

def view_txt(file_path):
    with open(file_path, 'r') as file:
        lines = file.readlines()
    
    data = []
    for line in lines:
        parts = line.strip().split()
        sequence_id = int(parts[0])
        sequence = list(map(lambda x: int(float(x)), parts[1:]))
        data.append([sequence_id, sequence])
    
    df = pd.DataFrame(data, columns=['sequence_num', 'sequence'])
    return df

def avg_seq_len(df):
    sequence_lengths = df['sequence'].apply(len)
    average_length = sequence_lengths.mean()
    return average_length

input = view_txt(sequence_output_path)
print(input.shape)
input.head()

result.contents_id = result.contents_id.astype(int)

# contents_id의 값을 문자열과 정수 형태로 변환
result_values = set(result['contents_id'].astype(str).values)

# 데이터 확인 : True면 다 있는 거
all(all(str(item) in result_values for item in row['sequence']) 
    for _, row in input.iterrows())

input

result.contents_id.nunique()
