==> 08-1. LLMEmb_embedding.ipynb <==
import os
import pickle
import subprocess
from tqdm import tqdm

import matplotlib.pyplot as plt
from sklearn.decomposition import PCA

import torch
from peft import LoraConfig, get_peft_model
from transformers import AutoTokenizer, AutoModelForCausalLM, BitsAndBytesConfig, TrainingArguments, pipeline, logging, Trainer, AutoConfig, LlamaConfig

# Huggingface에서 LLaMA 모델 권한 신청 이후, 권한 신청한 계정의 토큰을 [your token]에 입력
subprocess.run(["huggingface-cli", "login", "--token", "[your token]"])

def open_data_dict(file_path):
    with open(file_path, 'rb') as file:
        contents_dict = pickle.load(file)
        contents_dict = {int(key): value for key, value in contents_dict.items()}
        # contents_dict = {int(key.split('.')[0]): value for key, value in contents_dict.items()}

    return contents_dict

input_dict = open_data_dict('/data/log-data-2024/20241127_Final/content_meta_dict_20241127.pickle')
input_dict

# 모델 토크나이저 설정
model_ckpt = "meta-llama/Meta-Llama-3.1-8B"
tokenizer = AutoTokenizer.from_pretrained(model_ckpt)

# 토큰 길이 계산
token_lengths = []

for key, text in input_dict.items():
    tokens = tokenizer(text, truncation=False)['input_ids']  # 토큰화
    token_lengths.append(len(tokens))  # 토큰 길이 추가

# 토큰 길이 분포 시각화
plt.figure(figsize=(10, 6))
plt.hist(token_lengths, bins=30, color='blue', alpha=0.7)
plt.title("Distribution of Token Lengths")
plt.xlabel("Token Length")
plt.ylabel("Frequency")
plt.grid(True)
plt.show()

# 모델 및 토크나이저 설정
model_ckpt = "meta-llama/Meta-Llama-3.1-8B"
torch_dtype = torch.float16

bnb_config = BitsAndBytesConfig(
    bnb_4bit_quant_type="nf4",
    bnb_4bit_compute_dtype=torch_dtype,
    bnb_4bit_use_double_quant=False
)

tokenizer = AutoTokenizer.from_pretrained(model_ckpt)
tokenizer.pad_token = tokenizer.eos_token

model = AutoModelForCausalLM.from_pretrained(
    model_ckpt,
    low_cpu_mem_usage=True,
    quantization_config=bnb_config,
    torch_dtype=torch_dtype,
    device_map='auto'
)

model.config.use_cache = False

peft_config = LoraConfig(
    r=16,
    lora_alpha=32,
    lora_dropout=0.1,
    bias="none",
    task_type="CAUSAL_LM"
)
peft_config.inference_mode = False
model = get_peft_model(model, peft_config)

# Dictionary를 텍스트로 변환
embeddings_dict = {}

# 각 키에 대해 임베딩 계산
for key, text in tqdm(input_dict.items(), desc="Processing Texts", unit="entry"):
    inputs = tokenizer(text, return_tensors="pt", padding=True, truncation=False)
    with torch.no_grad():
        outputs = model(**inputs, output_hidden_states=True)
        embeddings = outputs.hidden_states[-1]  # 마지막 히든 레이어를 임베딩으로 사용 (torch.Size([1, N, 4096]), N은 토큰의 수)
        mean_embeddings = embeddings.mean(dim=1)  # [1, N, 4096] -> [1, 4096]로 평균화
        embeddings_dict[key] = mean_embeddings.cpu().numpy()  # 키와 임베딩을 dictionary에 저장
        embeddings_dict[key] = mean_embeddings.to(dtype=torch.float16)  # 키와 임베딩을 torch.float16 형태로 저장

embeddings_dict

# 패딩을 하는 '0' 토큰에 대한 임베딩을 추가
def add_zero_tensor(embeddings_dict):
    zero_tensor = torch.zeros((1, 4096), dtype=torch.float16)
    embeddings_dict[0] = zero_tensor
    return embeddings_dict

embeddings_dict = add_zero_tensor(embeddings_dict)
print(embeddings_dict[0])
print(embeddings_dict[0].size())

print(len(embeddings_dict))

# 결과를 .pt 파일로 저장
output_file_path = './data/embedding_vec_dict_241127.pt'
torch.save(embeddings_dict, output_file_path)

# PCA를 사용하여 임베딩 차원 축소
embedding_matrix = torch.stack([embeddings_dict[key].squeeze(0) for key in embeddings_dict.keys()])
embedding_matrix = embedding_matrix.view(embedding_matrix.size(0), -1)

pca = PCA(n_components=128)  # or 512, 256, 128 ()
reduced_embeddings = pca.fit_transform(embedding_matrix.cpu().numpy())
reduced_embeddings = torch.tensor(reduced_embeddings, dtype=torch.float16)
embedding_vec_dict_reduced = {key: reduced_embeddings[i] for i, key in enumerate(embeddings_dict.keys())}

torch.save(embedding_vec_dict_reduced, './data/embedding_vec_dict_241127_reduced_128.pt')
