==> 08-2. LLMEmb_training.ipynb <==
# 모델 훈련 전 확인해야 할 사항

# model/_abstract_model.py

## 1. self.embedding_vec_dict에 임베딩 벡터 위치와 일치하는지 확인


# main.py

## 1. 데이터 위치: data_dir, data_name 지정
## 2. 모델 저장 이름: train_name 지정
## 3. 임베딩 벡터 크기: hidden_size와 일치하는지 확인

# trainers.py

## 1. cuda gpu 사용해서 훈련할 것인지 아닌지 확인

!python main.py
